fix: ytm segment pull request fixes
