+++ httpclient.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"net/http"
 )
 
@@ -14,13 +13,3 @@ type httpClient interface {
 var (
 	client httpClient = &http.Client{}
 )
-
-// Get an HTTP response by sending an HTTP GET request to the specified URL.
-func Get(url string, headers http.Header) (*http.Response, error) {
-	request, err := http.NewRequestWithContext(context.Background(), http.MethodGet, url, nil)
-	if err != nil {
-		return nil, err
-	}
-	request.Header = headers
-	return client.Do(request)
-}
